requirements.txt
appdirs==1.4.4
aspy.yaml==1.3.0
attrs==20.3.0
beautifulsoup4==4.9.3
black==19.3b0
bs4==0.0.1
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
distlib==0.3.1
fake-useragent==0.1.11
filelock==3.0.12
future==0.16.0
identify==2.2.4
idna==2.7
importlib-metadata==4.0.1
lxml==4.6.2
nodeenv==1.6.0
parse==1.18.0
pre-commit==1.16.1
pyee==7.0.4
pyppeteer==0.2.3
pyquery==1.4.3
python-dotenv==0.17.1
python-slugify==3.0.2
PyYAML==5.4.1
requests==2.20.1
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
text-unidecode==1.2
toml==0.10.2
tqdm==4.55.1
urllib3==1.24.2
virtualenv==20.4.4
w3lib==1.22.0
webdriver-manager==3.4.1
websockets==8.1
wget==3.2
zipp==3.4.1
